+++ bot.py
@@ -60,8 +60,6 @@ if not TOKEN:
 
 intents = discord.Intents.default()
 intents.guilds = True
-intents.members = True
-intents.presences = True
 intents.messages = True
 intents.message_content = True
 intents.reactions = True